+++ packages/pn-personafisica-webapp/src/components/Contacts/DigitalContactManagement.tsx
@@ -79,8 +79,8 @@ const DigitalContactManagement: React.FC = () => {
       }
       activeStep={activeStep}
       setActiveStep={setActiveStep}
-      onExit={() => navigate(-1)}
       slots={{
+        exitButton: () => <></>,
         prevButton: getPreviouButton,
         nextButton:
           currentAction === DigitalDomicileManagementAction.ADD_SPECIAL_CONTACT