+++ packages/pn-personafisica-webapp/src/components/Contacts/EmailContactItem.tsx
@@ -1,8 +1,8 @@
 import { useRef, useState } from 'react';
-import { useTranslation } from 'react-i18next';
+import { Trans, useTranslation } from 'react-i18next';
 
 import PowerSettingsNewIcon from '@mui/icons-material/PowerSettingsNew';
-import { Button, Chip, Typography } from '@mui/material';
+import { Button, Chip, Divider, Typography } from '@mui/material';
 import { PnInfoCard, appStateActions } from '@pagopa-pn/pn-commons';
 
 import { PFEventsType } from '../../models/PFEventsType';
@@ -28,7 +28,6 @@ import SpecialContacts from './SpecialContacts';
 
 enum ModalType {
   EXISTING = 'existing',
-  DISCLAIMER = 'disclaimer',
   CODE = 'code',
   DELETE = 'delete',
   INFORMATIVE = 'informative',
@@ -179,6 +178,40 @@ const EmailContactItem: React.FC = () => {
     return 'default';
   };
 
+  const getRemoveModalTitle = () => {
+    if (blockDelete) {
+      return t(`courtesy-contacts.block-remove-email-title`, { ns: 'recapiti' });
+    }
+    if (defaultSERCQ_SENDAddress) {
+      return t(`courtesy-contacts.remove-email-title-dod-enabled`, {
+        ns: 'recapiti',
+      });
+    }
+    return t(`courtesy-contacts.remove-email`, { ns: 'recapiti' });
+  };
+
+  const getRemoveModalMessage = () => {
+    if (blockDelete) {
+      return t(`courtesy-contacts.block-remove-email-message`, { ns: 'recapiti' });
+    }
+    if (defaultSERCQ_SENDAddress) {
+      return (
+        <Trans
+          i18nKey={'courtesy-contacts.remove-address-message-dod-enabled'}
+          ns={'recapiti'}
+          components={[
+            <Typography variant="body2" fontSize={'18px'} key={'paragraph1'} sx={{ mb: 2 }} />,
+            <Typography variant="body2" fontSize={'18px'} key={'paragraph2'} />,
+          ]}
+        />
+      );
+    }
+    return t(`courtesy-contacts.remove-email-message`, {
+      value: currentAddress.current.value,
+      ns: 'recapiti',
+    });
+  };
+
   const getActions = () =>
     isEmailActive
       ? [
@@ -257,7 +290,12 @@ const EmailContactItem: React.FC = () => {
           {t('courtesy-contacts.email-filled-description', { ns: 'recapiti' })}
         </Typography>
       )}
-      {!defaultSMSAddress && <SmsContactItem />}
+      {!defaultSMSAddress && (
+        <>
+          <Divider sx={{ mt: 3, mb: 3 }} />
+          <SmsContactItem />
+        </>
+      )}
       {showSpecialContactsSection && (
         <SpecialContacts
           addressType={AddressType.COURTESY}
@@ -281,15 +319,22 @@ const EmailContactItem: React.FC = () => {
       />
       <DeleteDialog
         showModal={modalOpen === ModalType.DELETE}
-        removeModalTitle={t(`courtesy-contacts.${blockDelete ? 'block-' : ''}remove-email-title`, {
-          ns: 'recapiti',
-        })}
-        removeModalBody={t(`courtesy-contacts.${blockDelete ? 'block-' : ''}remove-email-message`, {
-          value: currentAddress.current.value,
-          ns: 'recapiti',
-        })}
+        removeModalTitle={getRemoveModalTitle()}
+        removeModalBody={getRemoveModalMessage()}
         handleModalClose={() => setModalOpen(null)}
         confirmHandler={deleteConfirmHandler}
+        slotsProps={{
+          primaryButton: {
+            onClick: defaultSERCQ_SENDAddress ? () => setModalOpen(null) : deleteConfirmHandler,
+            label: defaultSERCQ_SENDAddress ? t('button.annulla') : undefined,
+          },
+          secondaryButton: {
+            onClick: defaultSERCQ_SENDAddress ? deleteConfirmHandler : () => setModalOpen(null),
+            label: defaultSERCQ_SENDAddress
+              ? t('courtesy-contacts.remove-email', { ns: 'recapiti' })
+              : undefined,
+          },
+        }}
         blockDelete={blockDelete}
       />
       <InformativeDialog