+++ packages/pn-personafisica-webapp/src/components/Contacts/__test__/DigitalContactActivation.test.tsx
@@ -1,18 +1,18 @@
 import MockAdapter from 'axios-mock-adapter';
 import { vi } from 'vitest';
 
-import { ConsentType, SERCQ_SEND_VALUE } from '@pagopa-pn/pn-commons';
+import { SERCQ_SEND_VALUE } from '@pagopa-pn/pn-commons';
 import { getById } from '@pagopa-pn/pn-commons/src/test-utils';
 
 import {
-  acceptTosPrivacyConsentBodyMock,
-  sercqSendTosPrivacyConsentMock,
+  acceptTosSercqSendBodyMock,
+  sercqSendTosConsentMock,
 } from '../../../__mocks__/Consents.mock';
 import { digitalAddressesSercq, digitalLegalAddresses } from '../../../__mocks__/Contacts.mock';
-import { fireEvent, render, screen, waitFor, within } from '../../../__test__/test-utils';
+import { fireEvent, render, waitFor, within } from '../../../__test__/test-utils';
 import { apiClient } from '../../../api/apiClients';
 import { AddressType, ChannelType, IOAllowedValues } from '../../../models/contacts';
-import { RECAPITI } from '../../../navigation/routes.const';
+import { NOTIFICHE } from '../../../navigation/routes.const';
 import DigitalContactActivation from '../DigitalContactActivation';
 import { fillCodeDialog } from './test-utils';
 
@@ -24,6 +24,7 @@ vi.mock('react-router-dom', async () => ({
 }));
 
 describe('DigitalContactActivation', () => {
+  const labelPrefix = 'legal-contacts.sercq-send-wizard';
   let mock: MockAdapter;
 
   beforeAll(() => {
@@ -44,7 +45,7 @@ describe('DigitalContactActivation', () => {
 
   it('render component', () => {
     const { getByText } = render(<DigitalContactActivation />);
-    const title = getByText('legal-contacts.sercq-send-wizard.title');
+    const title = getByText(`${labelPrefix}.title`);
     expect(title).toBeInTheDocument();
   });
 
@@ -85,26 +86,13 @@ describe('DigitalContactActivation', () => {
     const { getByTestId } = render(<DigitalContactActivation />);
     const stepper = getByTestId('desktopWizardStepper');
     expect(stepper).toBeInTheDocument();
-    const step1Label = within(stepper).getByText('legal-contacts.sercq-send-wizard.step_1.title');
+    const step1Label = within(stepper).getByText(`${labelPrefix}.step_1.title`);
     expect(step1Label).toBeInTheDocument();
     const sercqSendContent = getByTestId('sercqSendContactWizard');
     expect(sercqSendContent).toBeInTheDocument();
   });
 
   it('renders the second step label if has appIO and is disabled', async () => {
-    mock
-      .onPost('/bff/v1/addresses/LEGAL/default/SERCQ_SEND', {
-        value: SERCQ_SEND_VALUE,
-      })
-      .reply(204);
-    mock.onGet(/\/bff\/v2\/tos-privacy.*/).reply(200, sercqSendTosPrivacyConsentMock(false, false));
-    mock
-      .onPut(
-        '/bff/v2/tos-privacy',
-        acceptTosPrivacyConsentBodyMock(ConsentType.TOS_SERCQ, ConsentType.DATAPRIVACY_SERCQ)
-      )
-      .reply(200);
-
     const { getByTestId } = render(<DigitalContactActivation />, {
       preloadedState: {
         contactsState: {
@@ -121,15 +109,11 @@ describe('DigitalContactActivation', () => {
     });
     const stepper = getByTestId('desktopWizardStepper');
     expect(stepper).toBeInTheDocument();
-    const step2Label = within(stepper).getByText('legal-contacts.sercq-send-wizard.step_2.title');
+    const step2Label = within(stepper).getByText(`${labelPrefix}.step_2.title`);
     expect(step2Label).toBeInTheDocument();
 
-    const activateSercqSendButton = getByTestId('activateButton');
-    fireEvent.click(activateSercqSendButton);
-
-    await waitFor(() => {
-      expect(mock.history.post.length).toBe(1);
-    });
+    const howItWorksContinueButton = getByTestId('continueButton');
+    fireEvent.click(howItWorksContinueButton);
 
     const ioContactWizard = getByTestId('ioContactWizard');
     expect(ioContactWizard).toBeInTheDocument();
@@ -137,7 +121,7 @@ describe('DigitalContactActivation', () => {
 
   it('does not render the second step label if has no app IO contact', () => {
     const { queryByText } = render(<DigitalContactActivation />);
-    const step2Label = queryByText('legal-contacts.sercq-send-wizard.step_2.title');
+    const step2Label = queryByText(`${labelPrefix}.step_2.title`);
     expect(step2Label).not.toBeInTheDocument();
   });
 
@@ -156,11 +140,11 @@ describe('DigitalContactActivation', () => {
         },
       },
     });
-    const step2Label = queryByText('legal-contacts.sercq-send-wizard.step_2.title');
+    const step2Label = queryByText(`${labelPrefix}.step_2.title`);
     expect(step2Label).not.toBeInTheDocument();
   });
 
-  it('renders the second and third step labels if has no courtesy contacts', () => {
+  it('does render the second and third step labels if has no courtesy contacts', () => {
     const { queryByText } = render(<DigitalContactActivation />, {
       preloadedState: {
         contactsState: {
@@ -175,13 +159,13 @@ describe('DigitalContactActivation', () => {
         },
       },
     });
-    const step2Label = queryByText('legal-contacts.sercq-send-wizard.step_2.title');
+    const step2Label = queryByText(`${labelPrefix}.step_2.title`);
     expect(step2Label).toBeInTheDocument();
-    const step3Label = queryByText('legal-contacts.sercq-send-wizard.step_3.step-title');
+    const step3Label = queryByText(`${labelPrefix}.step_3.step-title`);
     expect(step3Label).toBeInTheDocument();
   });
 
-  it('does not render the third step label if has an active email address', () => {
+  it('does render the third step label if has an active email address', () => {
     const { queryByText } = render(<DigitalContactActivation />, {
       preloadedState: {
         contactsState: {
@@ -196,99 +180,32 @@ describe('DigitalContactActivation', () => {
         },
       },
     });
-    const step3Label = queryByText('legal-contacts.sercq-send-wizard.step_3.step-title');
-    expect(step3Label).not.toBeInTheDocument();
+    const step3Label = queryByText(`${labelPrefix}.step_3.step-title`);
+    expect(step3Label).toBeInTheDocument();
   });
 
-  it('shows the confirmation modals trying to skip AppIO and email activation', async () => {
-    mock
-      .onPost('/bff/v1/addresses/LEGAL/default/SERCQ_SEND', {
-        value: SERCQ_SEND_VALUE,
-      })
-      .reply(204);
-    mock.onGet(/\/bff\/v2\/tos-privacy.*/).reply(200, sercqSendTosPrivacyConsentMock(false, false));
-    mock
-      .onPut(
-        '/bff/v2/tos-privacy',
-        acceptTosPrivacyConsentBodyMock(ConsentType.TOS_SERCQ, ConsentType.DATAPRIVACY_SERCQ)
-      )
-      .reply(200);
-
-    const { getByRole, getByTestId, getByText } = render(<DigitalContactActivation />, {
+  it('does render the third step label if has an active phone address', () => {
+    const { queryByText } = render(<DigitalContactActivation />, {
       preloadedState: {
         contactsState: {
           digitalAddresses: [
             {
               addressType: AddressType.COURTESY,
               senderId: 'default',
-              channelType: ChannelType.IOMSG,
-              value: IOAllowedValues.DISABLED,
+              channelType: ChannelType.SMS,
+              value: '+39333123456',
             },
           ],
         },
       },
     });
-    const activateSercqSendButton = getByTestId('activateButton');
-    fireEvent.click(activateSercqSendButton);
-
-    await waitFor(() => {
-      expect(mock.history.post.length).toBe(1);
-    });
-
-    // IO Step
-    const ioSkipButton = getByRole('button', { name: 'button.not-now' });
-    fireEvent.click(ioSkipButton);
-
-    let dialog = await waitFor(() => screen.getByRole('dialog'));
-    expect(dialog).toBeInTheDocument();
-
-    getByText('courtesy-contacts.confirmation-modal-title');
-    getByText('courtesy-contacts.confirmation-modal-io-content');
-    getByText('courtesy-contacts.confirmation-modal-io-accept');
-    const ioConfirmSkipButton = getByText('button.do-later');
-
-    fireEvent.click(ioConfirmSkipButton);
-    await waitFor(() => {
-      expect(dialog).not.toBeInTheDocument();
-    });
-
-    // Email Step
-    getByText('legal-contacts.sercq-send-wizard.step_3.title');
-    const mailSkipButton = getByRole('button', { name: 'button.not-now' });
-    fireEvent.click(mailSkipButton);
-
-    dialog = await waitFor(() => screen.getByRole('dialog'));
-    expect(dialog).toBeInTheDocument();
-
-    getByText('courtesy-contacts.confirmation-modal-title');
-    getByText('courtesy-contacts.confirmation-modal-email-content');
-    getByText('courtesy-contacts.confirmation-modal-email-accept');
-    const mailConfirmSkipButton = getByText('button.do-later');
-
-    fireEvent.click(mailConfirmSkipButton);
-    await waitFor(() => {
-      expect(dialog).not.toBeInTheDocument();
-    });
-
-    // Thank-you page
-    const feedbackStep = getByTestId('wizard-feedback-step');
-    expect(feedbackStep).toBeInTheDocument();
-
-    const feedbackTitle = getByTestId('wizard-feedback-title');
-    expect(feedbackTitle).toHaveTextContent(
-      'legal-contacts.sercq-send-wizard.feedback.title-activation'
-    );
-    const feedbackButton = getByTestId('wizard-feedback-button');
-    expect(feedbackButton).toHaveTextContent(
-      'legal-contacts.sercq-send-wizard.feedback.go-to-contacts'
-    );
-    fireEvent.click(feedbackButton);
-    expect(mockNavigateFn).toHaveBeenCalledTimes(1);
-    expect(mockNavigateFn).toHaveBeenCalledWith(RECAPITI);
+    const step3Label = queryByText(`${labelPrefix}.step_3.step-title`);
+    expect(step3Label).toBeInTheDocument();
   });
 
-  it('adds an email correctly (verify skip and confirm buttons)', async () => {
+  it('adds an email correctly', async () => {
     const mailValue = 'nome.utente@mail.it';
+    // mock new email api calls
     mock.onPost('/bff/v1/addresses/COURTESY/default/EMAIL', { value: mailValue }).reply(200, {
       result: 'CODE_VERIFICATION_REQUIRED',
     });
@@ -299,18 +216,14 @@ describe('DigitalContactActivation', () => {
       })
       .reply(204);
 
+    // mock SERCQ activation api call
     mock
       .onPost('/bff/v1/addresses/LEGAL/default/SERCQ_SEND', {
         value: SERCQ_SEND_VALUE,
       })
       .reply(204);
-    mock.onGet(/\/bff\/v2\/tos-privacy.*/).reply(200, sercqSendTosPrivacyConsentMock(false, false));
-    mock
-      .onPut(
-        '/bff/v2/tos-privacy',
-        acceptTosPrivacyConsentBodyMock(ConsentType.TOS_SERCQ, ConsentType.DATAPRIVACY_SERCQ)
-      )
-      .reply(200);
+    mock.onGet(/\/bff\/v2\/tos-privacy.*/).reply(200, sercqSendTosConsentMock(false));
+    mock.onPut('/bff/v2/tos-privacy', acceptTosSercqSendBodyMock).reply(200);
 
     const result = render(<DigitalContactActivation />, {
       preloadedState: {
@@ -326,15 +239,28 @@ describe('DigitalContactActivation', () => {
         },
       },
     });
-    const activateSercqSendButton = result.getByTestId('activateButton');
-    fireEvent.click(activateSercqSendButton);
+    const howItWorksContinueButton = result.getByTestId('continueButton');
+    fireEvent.click(howItWorksContinueButton);
+
+    let emailSmsContinueButton = result.getByText('button.continue');
+    expect(emailSmsContinueButton).toBeInTheDocument();
+    expect(emailSmsContinueButton).toBeEnabled();
+
+    fireEvent.click(emailSmsContinueButton);
+
+    const confirmationDialog = result.getByRole('dialog');
+
+    expect(confirmationDialog).toBeInTheDocument();
+    within(confirmationDialog).getByText('courtesy-contacts.confirmation-modal-title');
+    within(confirmationDialog).getByText('courtesy-contacts.confirmation-modal-email-content');
+
+    const dismissButton = result.getByRole('button', { name: 'button.understand' });
+
+    fireEvent.click(dismissButton);
 
     await waitFor(() => {
-      expect(mock.history.post.length).toBe(1);
+      expect(confirmationDialog).not.toBeInTheDocument();
     });
-
-    expect(result.getByText('button.not-now')).toBeInTheDocument();
-    expect(result.queryByText('button.conferma')).not.toBeInTheDocument();
 
     // insert new email
     const form = result.container.querySelector('form');
@@ -343,6 +269,13 @@ describe('DigitalContactActivation', () => {
     await waitFor(() => expect(input).toHaveValue(mailValue));
     const button = result.getByTestId('default_email-button');
     fireEvent.click(button);
+
+    // informative dialog
+    await waitFor(() => {
+      expect(result.getByRole('dialog')).toBeInTheDocument();
+    });
+    const infoButton = result.getByRole('button', { name: 'button.understand' });
+    fireEvent.click(infoButton);
 
     // inser otp and confirm
     const dialog = await fillCodeDialog(result);
@@ -353,83 +286,117 @@ describe('DigitalContactActivation', () => {
     expect(emailValue).toBeInTheDocument();
     expect(emailValue).toHaveTextContent(mailValue);
 
-    const confirmButton = result.getByText('button.conferma');
-    expect(confirmButton).toBeInTheDocument();
-    expect(result.queryByText('button.not-now')).not.toBeInTheDocument();
-    fireEvent.click(confirmButton);
+    emailSmsContinueButton = result.getByText('button.continue');
+    expect(emailSmsContinueButton).toBeInTheDocument();
 
-    // thankyou page
-    const feedbackStep = result.getByTestId('wizard-feedback-step');
-    expect(feedbackStep).toBeInTheDocument();
+    fireEvent.click(emailSmsContinueButton);
+
+    // recap step: check disclaimer and press button to enable sercq
+    const enableSercqStepTitle = result.getByText(`${labelPrefix}.step_4.title`);
+    expect(enableSercqStepTitle).toBeInTheDocument();
+
+    const disclaimerCkb = getById(result.container, 'disclaimer');
+    expect(disclaimerCkb).not.toBeChecked();
+
+    fireEvent.click(disclaimerCkb);
+
+    const enableSercqButton = result.getByTestId('activateButton');
+    fireEvent.click(enableSercqButton);
+
+    // feedback step
+    await waitFor(() => {
+      const feedbackStep = result.getByTestId('wizard-feedback-step');
+      expect(feedbackStep).toBeInTheDocument();
+    });
 
     const feedbackTitle = result.getByTestId('wizard-feedback-title');
-    expect(feedbackTitle).toHaveTextContent(
-      'legal-contacts.sercq-send-wizard.feedback.title-activation'
-    );
+    expect(feedbackTitle).toHaveTextContent(`${labelPrefix}.feedback.title-sercq_send-activation`);
+
+    const feedbackContent = result.getByTestId('wizard-feedback-content');
+    expect(feedbackContent).toHaveTextContent(`${labelPrefix}.feedback.content-sercq_send`);
+
     const feedbackButton = result.getByTestId('wizard-feedback-button');
-    expect(feedbackButton).toHaveTextContent(
-      'legal-contacts.sercq-send-wizard.feedback.go-to-contacts'
-    );
+    expect(feedbackButton).toHaveTextContent('button.understand');
     fireEvent.click(feedbackButton);
     expect(mockNavigateFn).toHaveBeenCalledTimes(1);
-    expect(mockNavigateFn).toHaveBeenCalledWith(RECAPITI);
+    expect(mockNavigateFn).toHaveBeenCalledWith(NOTIFICHE);
   });
 
   it('renders component correctly when transferring', async () => {
+    // mock SERCQ activation api call
     mock
       .onPost('/bff/v1/addresses/LEGAL/default/SERCQ_SEND', {
         value: SERCQ_SEND_VALUE,
       })
       .reply(204);
-    mock.onGet(/\/bff\/v2\/tos-privacy.*/).reply(200, sercqSendTosPrivacyConsentMock(false, false));
-    mock
-      .onPut(
-        '/bff/v2/tos-privacy',
-        acceptTosPrivacyConsentBodyMock(ConsentType.TOS_SERCQ, ConsentType.DATAPRIVACY_SERCQ)
-      )
-      .reply(200);
+    mock.onGet(/\/bff\/v2\/tos-privacy.*/).reply(200, sercqSendTosConsentMock(false));
+    mock.onPut('/bff/v2/tos-privacy', acceptTosSercqSendBodyMock).reply(200);
 
-    const { getByTestId, queryByTestId } = render(<DigitalContactActivation isTransferring />, {
-      preloadedState: {
-        contactsState: {
-          digitalAddresses: [
-            ...digitalLegalAddresses,
-            {
-              addressType: AddressType.COURTESY,
-              senderId: 'default',
-              channelType: ChannelType.EMAIL,
-              value: 'nome.utente@mail.it',
-            },
-            {
-              addressType: AddressType.COURTESY,
-              senderId: 'default',
-              channelType: ChannelType.IOMSG,
-              value: IOAllowedValues.ENABLED,
-            },
-          ],
+    const { container, getByTestId, queryByTestId, getByText } = render(
+      <DigitalContactActivation isTransferring />,
+      {
+        preloadedState: {
+          contactsState: {
+            digitalAddresses: [
+              ...digitalLegalAddresses,
+              {
+                addressType: AddressType.COURTESY,
+                senderId: 'default',
+                channelType: ChannelType.EMAIL,
+                value: 'nome.utente@mail.it',
+              },
+              {
+                addressType: AddressType.COURTESY,
+                senderId: 'default',
+                channelType: ChannelType.IOMSG,
+                value: IOAllowedValues.ENABLED,
+              },
+            ],
+          },
         },
-      },
-    });
+      }
+    );
 
     const pecSection = queryByTestId('pec-contact-wizard');
     expect(pecSection).not.toBeInTheDocument();
 
     const wizardTitle = getByTestId('wizard-title');
-    expect(wizardTitle).toHaveTextContent('legal-contacts.sercq-send-wizard.title-transfer');
+    expect(wizardTitle).toHaveTextContent(`${labelPrefix}.title-transfer`);
 
-    const activateSercqSendButton = getByTestId('activateButton');
-    fireEvent.click(activateSercqSendButton);
+    const howItWorksContinueButton = getByTestId('continueButton');
+    fireEvent.click(howItWorksContinueButton);
 
+    const emailSmsContinueButton = getByText('button.continue');
+    expect(emailSmsContinueButton).toBeInTheDocument();
+    expect(emailSmsContinueButton).toBeEnabled();
+    fireEvent.click(emailSmsContinueButton);
+
+    // recap step: check disclaimer and press button to enable sercq
+    const disclaimerCkb = getById(container, 'disclaimer');
+    expect(disclaimerCkb).not.toBeChecked();
+
+    fireEvent.click(disclaimerCkb);
+
+    const enableSercqButton = getByTestId('activateButton');
+    fireEvent.click(enableSercqButton);
+
+    // feedback step
     await waitFor(() => {
-      expect(mock.history.post.length).toBe(1);
+      const feedbackStep = getByTestId('wizard-feedback-step');
+      expect(feedbackStep).toBeInTheDocument();
     });
 
-    const feedbackStep = getByTestId('wizard-feedback-step');
-    expect(feedbackStep).toBeInTheDocument();
-
     const feedbackTitle = getByTestId('wizard-feedback-title');
-    expect(feedbackTitle).toHaveTextContent(
-      'legal-contacts.sercq-send-wizard.feedback.title-transfer'
-    );
+    expect(feedbackTitle).toHaveTextContent(`${labelPrefix}.feedback.title-sercq_send-transfer`);
+
+    const feedbackContent = getByTestId('wizard-feedback-content');
+    expect(feedbackContent).toHaveTextContent(`${labelPrefix}.feedback.content-sercq_send`);
+
+    const feedbackButton = getByTestId('wizard-feedback-button');
+    expect(feedbackButton).toHaveTextContent('button.understand');
+    fireEvent.click(feedbackButton);
+
+    expect(mockNavigateFn).toHaveBeenCalledTimes(1);
+    expect(mockNavigateFn).toHaveBeenCalledWith(NOTIFICHE);
   });
 });