+++ packages/pn-personagiuridica-webapp/src/components/Contacts/PecContactWizard.tsx
@@ -6,6 +6,7 @@ import * as yup from 'yup';
 
 import {
   Alert,
+  Button,
   Checkbox,
   FormControl,
   FormControlLabel,
@@ -17,7 +18,7 @@ import { PnWizard, PnWizardStep } from '@pagopa-pn/pn-commons';
 import { ButtonNaked } from '@pagopa/mui-italia';
 
 import { AddressType, ChannelType, SaveDigitalAddressParams } from '../../models/contacts';
-import { RECAPITI } from '../../navigation/routes.const';
+import { NOTIFICHE } from '../../navigation/routes.const';
 import { createOrUpdateAddress } from '../../redux/contact/actions';
 import { contactsSelectors } from '../../redux/contact/reducers';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
@@ -43,6 +44,11 @@ const PecContactWizard: React.FC<Props> = ({
   const { defaultSERCQ_SENDAddress } = useAppSelector(contactsSelectors.selectAddresses);
   const [openCodeModal, setOpenCodeModal] = useState(false);
   const { IS_DOD_ENABLED } = getConfiguration();
+
+  const feedbackTitleLabel = `legal-contacts.sercq-send-wizard.feedback.title-pec-${
+    isTransferring ? 'transfer' : 'activation'
+  }`;
+  const feedbackContentLabel = 'legal-contacts.sercq-send-wizard.feedback.content-pec';
 
   const validationSchema = yup.object().shape({
     pec: pecValidationSchema(t),
@@ -83,7 +89,7 @@ const PecContactWizard: React.FC<Props> = ({
           return;
         }
         setOpenCodeModal(false);
-        return isTransferring ? setActiveStep(activeStep + 1) : navigate(RECAPITI);
+        return setActiveStep(activeStep + 1);
       })
       .catch(() => {});
   };
@@ -106,8 +112,9 @@ const PecContactWizard: React.FC<Props> = ({
         }
         activeStep={activeStep}
         setActiveStep={setActiveStep}
-        onExit={() => navigate(-1)}
         slots={{
+          exitButton: () => <></>,
+          nextButton: () => <></>,
           prevButton: () => (
             <ButtonNaked
               onClick={handlePreviousBtnClick}
@@ -121,23 +128,16 @@ const PecContactWizard: React.FC<Props> = ({
           ),
         }}
         slotsProps={{
-          nextButton: {
-            onClick: () => formik.submitForm(),
-          },
           container: {
             'data-testid': 'pec-contact-wizard',
           },
-          feedback: isTransferring
-            ? {
-                title: t(
-                  `legal-contacts.sercq-send-wizard.feedback.title-${
-                    isTransferring ? 'transfer' : 'activation'
-                  }`
-                ),
-                buttonText: t('legal-contacts.sercq-send-wizard.feedback.go-to-contacts'),
-                onClick: () => navigate(RECAPITI),
-              }
-            : undefined,
+          feedback: {
+            title: t(feedbackTitleLabel),
+            content: t(feedbackContentLabel),
+            buttonText: t('button.understand', { ns: 'common' }),
+            onClick: () => navigate(NOTIFICHE),
+          },
+          actions: { justifyContent: 'center' },
         }}
       >
         <PnWizardStep>
@@ -184,6 +184,11 @@ const PecContactWizard: React.FC<Props> = ({
             <FormControlLabel
               control={
                 <Checkbox
+                  sx={
+                    formik.touched.disclaimer && formik.errors.disclaimer
+                      ? { color: 'error.dark' }
+                      : { color: 'text.secondary' }
+                  }
                   name="disclaimer"
                   id="disclaimer"
                   required
@@ -204,6 +209,16 @@ const PecContactWizard: React.FC<Props> = ({
               </FormHelperText>
             )}
           </FormControl>
+          <Button
+            fullWidth
+            variant="contained"
+            color="primary"
+            onClick={formik.submitForm}
+            sx={{ mt: 3 }}
+            data-testid="next-button"
+          >
+            {t('legal-contacts.sercq-send-active-pec-enabled', { ns: 'recapiti' })}
+          </Button>
         </PnWizardStep>
       </PnWizard>