+++ packages/pn-personagiuridica-webapp/src/components/Contacts/__test__/DeleteDialog.test.tsx
@@ -21,7 +21,6 @@ describe('DeleteDialog Component', () => {
         handleModalClose={closeHandler}
         removeModalTitle="remove-title"
         removeModalBody="remove-modal-body"
-        blockDelete={false}
         confirmHandler={confirmHandler}
       />
     );
@@ -75,7 +74,6 @@ describe('DeleteDialog Component', () => {
         handleModalClose={closeHandler}
         removeModalTitle="remove-title"
         removeModalBody="remove-modal-body"
-        blockDelete={false}
         confirmHandler={confirmHandler}
       />
     );